+++ src/bump_string.rs
@@ -285,6 +285,12 @@ impl<A: BumpAllocator> BumpString<A> {
         unsafe { transmute_value(self) }
     }
 
+    #[track_caller]
+    #[inline(always)]
+    pub(crate) fn assert_char_boundary(&self, index: usize) {
+        unsafe { self.fixed.cook_ref() }.assert_char_boundary(index);
+    }
+
     /// Splits the string into two by removing the specified range.
     ///
     /// This method does not allocate and does not change the order of the elements.
@@ -571,6 +577,26 @@ impl<A: BumpAllocator> BumpString<A> {
         // SAFETY: `BumpVec<u8>` and `BumpString` have the same representation;
         // only the invariant that the bytes are utf8 is different.
         transmute_mut(self)
+    }
+
+    /// Returns a raw pointer to the slice, or a dangling raw pointer
+    /// valid for zero sized reads.
+    #[inline]
+    #[must_use]
+    pub fn as_ptr(&self) -> *const u8 {
+        self.fixed.as_ptr()
+    }
+
+    /// Returns an unsafe mutable pointer to slice, or a dangling
+    /// raw pointer valid for zero sized reads.
+    #[inline]
+    pub fn as_mut_ptr(&mut self) -> *mut u8 {
+        self.fixed.as_mut_ptr()
+    }
+
+    #[inline(always)]
+    pub(crate) unsafe fn set_len(&mut self, new_len: usize) {
+        self.fixed.set_len(new_len);
     }
 
     error_behavior_generic_methods_allocation_failure! {
@@ -1071,6 +1097,83 @@ impl<A: BumpAllocator> BumpString<A> {
 
                 ptr.add(len).write_bytes(0, additional);
                 vec.set_len(len + additional);
+            }
+
+            Ok(())
+        }
+
+        /// Removes the specified range in the string,
+        /// and replaces it with the given string.
+        /// The given string doesn't need to be the same length as the range.
+        do panics
+        /// Panics if the starting point or end point do not lie on a [`char`]
+        /// boundary, or if they're out of bounds.
+        impl
+        do examples
+        /// ```
+        /// # use bump_scope::{ Bump, BumpString };
+        /// # let bump: Bump = Bump::new();
+        /// let mut s = BumpString::from_str_in("α is alpha, β is beta", &bump);
+        /// let beta_offset = s.find('β').unwrap_or(s.len());
+        ///
+        /// // Replace the range up until the β from the string
+        /// s.replace_range(..beta_offset, "Α is capital alpha; ");
+        /// assert_eq!(s, "Α is capital alpha; β is beta");
+        /// ```
+        for fn replace_range
+        do examples
+        /// ```
+        /// # use bump_scope::{ Bump, BumpString };
+        /// # let bump: Bump = Bump::try_new()?;
+        /// let mut s = BumpString::try_from_str_in("α is alpha, β is beta", &bump)?;
+        /// let beta_offset = s.find('β').unwrap_or(s.len());
+        ///
+        /// // Replace the range up until the β from the string
+        /// s.try_replace_range(..beta_offset, "Α is capital alpha; ")?;
+        /// assert_eq!(s, "Α is capital alpha; β is beta");
+        /// # Ok::<(), bump_scope::allocator_api2::alloc::AllocError>(())
+        /// ```
+        for fn try_replace_range
+        #[inline]
+        use fn generic_replace_range<{R}>(&mut self, range: R, replace_with: &str)
+        where {
+            R: RangeBounds<usize>,
+        } in {
+            let Range { start, end } = polyfill::slice::range(range, ..self.len());
+
+            self.assert_char_boundary(start);
+            self.assert_char_boundary(end);
+
+            let range_len = end - start;
+            let given_len = replace_with.len();
+
+            let additional_len = given_len.saturating_sub(range_len);
+            self.generic_reserve(additional_len)?;
+
+            // move the tail
+            if range_len != given_len {
+                unsafe {
+                    let src = self.as_ptr().add(end);
+                    let dst = self.as_mut_ptr().add(start + given_len);
+                    let len = self.len() - end;
+                    src.copy_to(dst, len);
+                }
+            }
+
+            // fill with given string
+            unsafe {
+                let src = replace_with.as_ptr();
+                let dst = self.as_mut_ptr().add(start);
+                let len = replace_with.len();
+                src.copy_to(dst, len);
+            }
+
+            // update len
+            unsafe {
+                // Casting to `isize` is fine because per `Layout`'s rules all the `*len`s must be
+                // less than isize::MAX. Subtracting two positive `isize`s can't overflow.
+                let len_diff = given_len as isize - range_len as isize;
+                self.set_len((self.len() as isize + len_diff) as usize);
             }
 
             Ok(())